Cleaning of the JSONREPOS folder added
