+++ cake/Program.cs
@@ -324,16 +324,17 @@ public sealed class AppsRunTask : FrostingTask<BuildContext>
             context.Log.Warning($"Skipping apps run");
             return;
         }
-        KillProcess(context,"Siemens.Simatic.PlcSim.Advanced.UserInterface");
+
+        AppsRunTaskHelpers.KillProcess(context,"Siemens.Simatic.PlcSim.Advanced.UserInterface");
 
         bool summaryResult = true;
 
-        var createResult = CreateLogFile(context, "app_test_result");
+        var createResult = AppsRunTaskHelpers.CreateLogFile(context, "app_test_result");
 
         if (createResult.Success)
         {
             string logFilePath = createResult.FilePath;
-            WriteResult(context, "AppName,PlcSim,PlcHw,PlcSw,DotnetBuild,DotnetRun", logFilePath);
+            AppsRunTaskHelpers.WriteResult(context, "AppName,PlcSim,PlcHw,PlcSw,DotnetBuild,DotnetRun", logFilePath);
 
             foreach (var library in context.Libraries)
             {
@@ -350,21 +351,23 @@ public sealed class AppsRunTask : FrostingTask<BuildContext>
                         context.Log.Information($"Starting the application: {appName}");
                         context.Log.Information($"File of the application: {appFile}");
                         context.Log.Information($"###################################################");
+                        AppsRunTaskHelpers.WriteResult(context, " ", logFilePath);
+                        AppsRunTaskHelpers.WriteResult(context, appName, logFilePath, appendToSameLine: true);
 
-                        WriteResult(context, " ", logFilePath);
-                        WriteResult(context, appName, logFilePath, appendToSameLine: true);
+                        // Cleanup JSONREPOS
+                        AppsRunTaskHelpers.DeleteJsonReposFolder(context, appFolder);
 
                         // Initialize PLC Sim instance
-                        InitializePlcSimInstance(context, appName);
+                        AppsRunTaskHelpers.InitializePlcSimInstance(context, appName);
 
                         // Overwrite security files
-                        OverwriteSecurityFiles(context,appFile, context.PlcName);
+                        AppsRunTaskHelpers.OverwriteSecurityFiles(context,appFile, context.PlcName);
 
                         // Build and load PLC
-                        BuildAndLoadPlc(context, appFile, appName, logFilePath, ref summaryResult);
+                        AppsRunTaskHelpers.BuildAndLoadPlc(context, appFile, appName, logFilePath, ref summaryResult);
 
                         // Build and start HMI
-                        BuildAndStartHmi(context, appFile, appName, logFilePath, ref summaryResult);
+                        AppsRunTaskHelpers.BuildAndStartHmi(context, appFile, appName, logFilePath, ref summaryResult);
                     }
                 }
             }
@@ -373,452 +376,10 @@ public sealed class AppsRunTask : FrostingTask<BuildContext>
         {
             Console.Error.WriteLine("Failed to create the log file.");
         }
-        KillProcess(context, "Siemens.Simatic.PlcSim.Advanced.UserInterface");
+
+        AppsRunTaskHelpers.KillProcess(context, "Siemens.Simatic.PlcSim.Advanced.UserInterface");
         context.Log.Information("Apps run done.");
     }
-
-    private static (bool Success, string FilePath) CreateLogFile(BuildContext context, string fileNamePrefix = "app_test_result")
-    {
-        string timestamp = DateTime.Now.ToString("yyyyMMdd_HHmmss"); // Generate the timestamp
-        string fileName = $"{fileNamePrefix}_{timestamp}.csv"; // Create the full file name
-        string appsRunResultFileName = Path.GetFullPath(Path.Combine(context.AppTestResultsDir, fileName));
-
-        // Create result directory
-        try
-        {
-            context.EnsureDirectoryExists(context.AppTestResultsDir);
-        }
-        catch (Exception ex)
-        {
-            context.Log.Error($"Failed to create the directory {context.AppTestResultsDir}: {ex.Message}");
-            return (false, null);
-        }
-        context.Log.Information($"Directory: {context.AppTestResultsDir} has been created.");
-
-        // Create result file
-        try
-        {
-            using (File.Create(appsRunResultFileName)) { }
-            context.Log.Information($"File '{appsRunResultFileName}' has been created.");
-            return (context.FileExists(appsRunResultFileName), appsRunResultFileName);
-        }
-        catch (Exception ex)
-        {
-            context.Log.Error($"Failed to create the file {appsRunResultFileName}: {ex.Message}");
-            return (false, null);
-        }
-    }
-
-    private static void WriteResult(BuildContext context,string textToWrite, string logFilePath, bool appendToSameLine = false)
-    {
-        // Check if the log file exists
-        if (!context.FileExists(logFilePath))
-        {
-            context.Log.Error($"The specified log file does not exist: {logFilePath}");
-            return;
-        }
-
-        try
-        {
-            if (appendToSameLine)
-            {
-                // Append to the same line without a newline
-                File.AppendAllText(logFilePath, textToWrite);
-            }
-            else
-            {
-                // Append text with a newline
-                File.AppendAllText(logFilePath, textToWrite + Environment.NewLine);
-            }
-
-            context.Log.Information($"Text successfully written to file '{logFilePath}'.");
-            context.Log.Information("File content:");
-
-            // Read and display the content of the log file
-            var fileContent = File.ReadAllText(logFilePath);
-            context.Log.Information(fileContent);
-        }
-        catch (Exception ex)
-        {
-            context.Log.Error($"An error occurred while writing to the file: {ex.Message}");
-        }
-    }
-
-    private static void KillProcess(BuildContext context,string processName)
-    {
-        try
-        {
-            // Get all processes by name
-            Process[] processes = Process.GetProcessesByName(processName);
-
-            if (processes.Length > 0)
-            {
-                foreach (var process in processes)
-                {
-                    try
-                    {
-                        // Attempt to kill the process
-                        process.Kill(true); // Forcefully terminate the process
-                        context.Log.Information($"Successfully terminated process: {process.ProcessName} (ID: {process.Id})");
-                    }
-                    catch (Exception ex)
-                    {
-                        context.Log.Error($"Failed to terminate process: {process.ProcessName} (ID: {process.Id}) - {ex.Message}");
-                    }
-                }
-            }
-            else
-            {
-                context.Log.Information($"No processes named '{processName}' were found.");
-            }
-        }
-        catch (Exception ex)
-        {
-            context.Log.Error($"An error occurred while attempting to find or kill processes: {ex.Message}");
-        }
-    }
-
-    private static void InitializePlcSimInstance(BuildContext context, string instanceName)
-    {
-        string plcSimSourceDir = context.SourceDirPlcSim;
-        string memoryCardPath = context.PlcSimVirtualMemoryCardLocation;
-
-        // Validate the memory card path
-        if (!context.DirectoryExists(memoryCardPath))
-        {
-            context.Log.Error($"The provided path for the virtual memory card does not exist: {memoryCardPath}");
-            return;
-        }
-
-        // Validate the instance name
-        if (string.IsNullOrWhiteSpace(instanceName))
-        {
-            context.Log.Error("The provided instance name is empty.");
-            return;
-        }
-
-        // Determine the instance path
-        string instancePath = Path.GetFullPath(Path.Combine(memoryCardPath, instanceName));
-
-        if (context.DirectoryExists(instancePath))
-        {
-            // Delete all files and subfolders
-            context.Log.Information($"The instance path {instancePath} already exists. It needs to be cleaned before.");
-            try
-            {
-                context.CleanDirectory(instancePath, new CleanDirectorySettings() { Force = true });
-            }
-            catch (Exception ex)
-            {
-                context.Log.Error($"Failed to clean the directory {instancePath}: {ex.Message}");
-            }
-            context.Log.Information($"Directory {instancePath} has been cleaned.");
-        }
-
-        // Create the instance directory
-        try
-        {
-            context.EnsureDirectoryExists(instancePath);
-        }
-        catch (Exception ex)
-        {
-            context.Log.Error($"Failed to create directory {instancePath}: {ex.Message}");
-            return;
-        }
-        context.Log.Information($"Directory: {instancePath} has been created.");
-
-        // Source directory for PLC Sim files
-        if (!context.DirectoryExists(plcSimSourceDir))
-        {
-            context.Log.Error($"Source directory for the plcsim {plcSimSourceDir} does not exist.");
-            return;
-        }
-
-        // Copy plcsim virtual memory card content
-        try
-        {
-            context.CopyDirectory(plcSimSourceDir, instancePath);
-        }
-        catch (Exception ex)
-        {
-            context.Log.Error($"Failed to copy plcsim virtual memory card content from: {plcSimSourceDir} to: {instancePath}: {ex.Message}");
-            return;
-        }
-        context.Log.Information($"Plcsim virtual memory card content has been copied succesfullt from: {plcSimSourceDir} to: {instancePath}");
-    }
-
-    private static void OverwriteSecurityFiles(BuildContext context, string appYamlFile, string plcName)
-    {
-        string sourceDirSecurityFiles = context.SourceDirSecurityFiles;
-
-        // Check application YAML file
-        if (string.IsNullOrWhiteSpace(appYamlFile))
-        {
-            context.Log.Error("The provided YAML of the application is empty.");
-            return;
-        }
-
-        if (!context.FileExists(appYamlFile))
-        {
-            context.Log.Error($"The provided application file does not exist: {appYamlFile}");
-            return;
-        }
-
-        string appFolder = Path.GetFullPath(Path.GetDirectoryName(appYamlFile));
-        if (string.IsNullOrWhiteSpace(appFolder) || !context.DirectoryExists(appFolder))
-        {
-            context.Log.Error($"The provided path for the application does not exist: {appFolder}");
-            return;
-        }
-
-
-        if (string.IsNullOrWhiteSpace(plcName))
-        {
-            context.Log.Error("The provided PLC name is empty.");
-            return;
-        }
-
-        // Check source directory
-        if (string.IsNullOrWhiteSpace(sourceDirSecurityFiles))
-        {
-            context.Log.Error("The provided directory name for the source directory of the security files is empty.");
-            return;
-        }
-
-        if (!context.DirectoryExists(sourceDirSecurityFiles))
-        {
-            context.Log.Error($"Source directory for the security file does not exist: {sourceDirSecurityFiles}");
-            return;
-        }
-
-        // Handle certification folder
-        string appCertFolder = Path.Combine(appFolder, "certs");
-        if (context.DirectoryExists(appCertFolder))
-        {
-            context.Log.Information($"The application certification folder {appCertFolder} already exists. It needs to be cleaned before.");
-            try
-            {
-                context.CleanDirectory(appCertFolder, new CleanDirectorySettings() { Force = true});
-                context.Log.Information($"The application certification folder {appCertFolder} has been cleaned.");
-            }
-            catch (Exception ex)
-            {
-                context.Log.Error($"Failed to clean the application certification folder {appCertFolder}: {ex.Message}");
-            }
-        }
-        else
-        {
-            try
-            {
-                context.EnsureDirectoryExists(appCertFolder);
-                context.Log.Information($"The application certification folder {appCertFolder} has been created.");
-            }
-            catch (Exception ex)
-            {
-                context.Log.Error($"Failed to create the application certification folder {appCertFolder}: {ex.Message}");
-            }
-        }
-
-        // Handle certification subfolder for the plc: 'plcName'
-        string appCertSubFolder = Path.GetFullPath(Path.Combine(appCertFolder, plcName));
-        try
-        {
-            context.EnsureDirectoryExists (appCertSubFolder);
-            context.Log.Information($"The application certification  subfolder {appCertSubFolder} has been created.");
-        }
-        catch (Exception ex)
-        {
-            context.Log.Error($"Failed to create the application certification folder {appCertSubFolder}: {ex.Message}");
-        }
-
-        // Handle certification file
-        string plcSourceCertificateFile = $"{plcName}.cer";
-        string plcSourceCertificate = Path.GetFullPath(Path.Combine(sourceDirSecurityFiles, plcSourceCertificateFile));
-        if (context.FileExists(plcSourceCertificate))
-        {
-            try
-            {
-                context.CopyFileToDirectory(plcSourceCertificate, appCertSubFolder);
-                context.Log.Information($"Certification file '{plcSourceCertificateFile}' has been copied to: {appCertSubFolder}");
-            }
-            catch (Exception ex)
-            {
-                context.Log.Error($"Failed to copy the certification file '{plcSourceCertificateFile}': {ex.Message}");
-            }
-        }
-        else
-        {
-            context.Log.Error($"Certification file '{plcSourceCertificateFile}' does not exist in the directory: {sourceDirSecurityFiles}");
-        }
-
-        // Handle hwc folder
-        string appHwcFolder = Path.GetFullPath(Path.Combine(appFolder, "hwc"));
-        if (!context.DirectoryExists(appHwcFolder))
-        {
-            context.Log.Error($"The application's hwc folder does not exist: {appHwcFolder}");
-            return;
-        }
-
-        // Handle hwc.gen subfolder
-        string appHwcGenFolder = Path.GetFullPath(Path.Combine(appHwcFolder, "hwc.gen"));
-        if (context.DirectoryExists(appHwcGenFolder))
-        {
-            context.Log.Information($"The application's hwc.gen subfolder {appHwcGenFolder} already exists. It needs to be cleaned before.");
-            try
-            {
-                context.CleanDirectory(appHwcGenFolder, new CleanDirectorySettings() { Force = true });
-                context.Log.Information($"The application's hwc.gen subfolder {appHwcGenFolder} has been cleaned.");
-            }
-            catch (Exception ex)
-            {
-                context.Log.Error($"Failed to clean the application's hwc.gen subfolder {appHwcGenFolder}: {ex.Message}");
-            }
-        }
-        else
-        {
-            try
-            {
-                context.EnsureDirectoryExists(appHwcGenFolder);
-                context.Log.Information($"The application's hwc.gen subfolder {appHwcGenFolder} has been created.");
-            }
-            catch (Exception ex)
-            {
-                context.Log.Error($"Failed to create the application's hwc.gen subfolder {appHwcGenFolder}: {ex.Message}");
-            }
-        }
-
-        // Handle secutity configuration file
-        string securityConfigFileName = $"{plcName}.SecurityConfiguration.json";
-        string securityConfigFile = Path.GetFullPath(Path.Combine(sourceDirSecurityFiles, securityConfigFileName));
-        if (context.FileExists(securityConfigFile))
-        {
-            try
-            {
-                context.CopyFileToDirectory(securityConfigFile, appHwcGenFolder);
-                context.Log.Information($"Security configuration file '{securityConfigFileName}' successfully copied to: {appHwcGenFolder}");
-            }
-            catch (Exception ex)
-            {
-                context.Log.Error($"Failed to copy the file: {ex.Message}");
-            }
-        }
-        else
-        {
-            context.Log.Information($"Security configuration file '{securityConfigFileName}' does not exist in the directory: {sourceDirSecurityFiles}");
-        }
-    }
-
-    private static void BuildAndLoadPlc(BuildContext context, string appYamlFile, string appName, string logFilePath, ref bool summaryResult)
-    {
-        string plcName = context.PlcName;
-        string plcIpAddress = context.PlcIpAddress;
-
-        // Validate application YAML file
-        if (string.IsNullOrWhiteSpace(appYamlFile))
-        {
-            context.Log.Error("The provided YAML of the application is empty.");
-            return;
-        }
-
-        if (!File.Exists(appYamlFile))
-        {
-            context.Log.Error($"The provided application file does not exist: {appYamlFile}");
-            return;
-        }
-
-        if (string.IsNullOrWhiteSpace(appName))
-        {
-            context.Log.Error("The provided application name is empty.");
-            return;
-        }
-
-        string appFolder = Path.GetDirectoryName(appYamlFile);
-        if (string.IsNullOrWhiteSpace(appFolder) || !Directory.Exists(appFolder))
-        {
-            context.Log.Error($"The provided path for the application does not exist: {appFolder}");
-            return;
-        }
-
-        // Run "apax install"
-        ApaxCmd.ApaxInstall(context,appFolder);
-
-        // Run "apax plcsim"
-        string plcSimResult = ApaxCmd.ApaxPlcSim(context, appFolder, ref summaryResult);
-        WriteResult(context, plcSimResult, logFilePath, appendToSameLine: true);
-
-        // Run "apax hwu"
-        string hwuResult = ApaxCmd.ApaxHwu(context, appFolder, ref summaryResult);
-        WriteResult(context, hwuResult, logFilePath, appendToSameLine: true);
-
-        // Run "apax swfd"
-        string swfdResult = ApaxCmd.ApaxSwfd(context, appFolder, ref summaryResult);
-        WriteResult(context, swfdResult, logFilePath, appendToSameLine: true);
-    }
-
-    public static void BuildAndStartHmi(BuildContext context, string appYamlFile, string appName, string logFilePath, ref bool summaryResult)
-    {
-        // Validate application YAML file
-        if (string.IsNullOrWhiteSpace(appYamlFile))
-        {
-            context.Log.Error("The provided YAML of the application is empty.");
-            return;
-        }
-
-        if (!context.FileExists(appYamlFile))
-        {
-            context.Log.Error($"The provided application file does not exist: {appYamlFile}");
-            return;
-        }
-
-        if (string.IsNullOrWhiteSpace(appName))
-        {
-            context.Log.Error("The provided application name is empty.");
-            return;
-        }
-
-        string appFolder = Path.GetFullPath(Path.GetDirectoryName(appYamlFile));
-        if (string.IsNullOrWhiteSpace(appFolder) || !context.DirectoryExists(appFolder))
-        {
-            context.Log.Error($"The provided path for the application does not exist: {appFolder}");
-            return;
-        }
-
-        // Recreate solution file by running the slngen script
-        string slnGenPath = Path.GetFullPath(Path.GetFullPath(Path.Combine (appFolder, "..", "./slngen.ps1")));
-        DotNetCmd.RunPowershellScript(context,slnGenPath, "");
-
-        // Clean solution
-        string solutionFile = Path.GetFullPath(Path.Combine(appFolder, "../this.sln"));
-        DotNetCmd.DotNetClean(context, solutionFile, "-c Debug");
-
-        // Build solution
-        string buildResult = DotNetCmd.DotNetBuildWithResult(context, solutionFile, "-c Debug",ref summaryResult);
-        WriteResult(context, buildResult, logFilePath, appendToSameLine: true );
-
-        // Get blazor projects
-        var blazorFiles = Directory.GetFiles(appFolder, "*.csproj", SearchOption.AllDirectories).Where(file => file.Contains("blazor")).ToList();
-
-        if (blazorFiles.Any())
-        {
-            foreach (var blazorFile in blazorFiles)
-            {
-                context.Log.Information($"Application 'blazor' file: {blazorFile}");
-
-                // Filter out libraries by checking for <PackageId> in the project file
-                string csprojContent = File.ReadAllText(blazorFile);
-                if (!csprojContent.Contains("<PackageId>"))
-                {
-                    string runResult = DotNetCmd.DotNetRunWithResult(context, blazorFile, "-c Debug --framework net9.0",60 , ref summaryResult);
-                    WriteResult(context, runResult, logFilePath, appendToSameLine: true);
-                }
-            }
-        }
-        else
-        {
-            context.Log.Information("No files containing 'blazor' in the filename and ending with '.csproj' were found.");
-        }
-    }
-
 }
 
 [TaskName("CreateArtifacts")]
